+++ src/Extract/index.js
@@ -10,3 +10,4 @@
 import 'extractComponentNode';
 import 'extractRenderNode';
 import 'extractVirtualNode';
+import 'extractFunctionNode';